加了add和update檔案
